basic refacter, added typography.scss
