+++ troposphere/static/css/app/app.scss
@@ -60,8 +60,8 @@ $icon-font-path: "~bootstrap-sass/assets/fonts/bootstrap/";
     left: 0;
   }
   .modal-body {
-    padding-bottom: 75px;
-    min-height: 600px;
+    padding-bottom: 60px;
+    min-height: 500px;
   }
 }
 //
@@ -70,6 +70,7 @@ $icon-font-path: "~bootstrap-sass/assets/fonts/bootstrap/";
 //
 @import "variables";
 @import "layout";
+@import "typography";
 @import "common";
 @import "header";
 @import "dashboard";
