+++ troposphere/static/css/app/typography.scss
@@ -1,0 +1,61 @@
+body {
+    font-family:'Roboto' , sans-serif;
+}
+
+
+.display-4 {
+    font-size: 112px;
+    font-weight: 300;
+}
+
+.display-3 {
+    font-size: 56px;
+    font-weight: 400px;
+}
+
+.display-2 {
+    font-size: 45px;
+    font-weight: 400;
+}
+
+.display-1 {
+    font-size: 34px;
+    font-weight: 400;
+}
+
+.headline {
+    font-size: 24px;
+    font-weight: 500;
+    margin:0 0 10px
+}
+
+.title {
+    font-size: 20px;
+    font-weight: 500;
+    margin:0 0 10px;
+}
+
+.subheading {
+    font-size: 16px;
+    font-weight: 400;
+}
+
+.body-2 {
+    font-size: 14px;
+    font-weight: 500;
+}
+
+.body-1 {
+    font-size: 14px;
+    font-weight: 400;
+}
+
+.caption { 
+    font-size: 12px;
+    font-weight: 400;
+}
+
+
+hr {
+    margin: 0 0 20px;
+}
