+++ troposphere/static/css/app/variables.scss
@@ -13,3 +13,4 @@ $grey--xxlight: #F7F8F8;
 //Shadow Effects
 
 $shadow-1: 0 1px 5px rgba(43,59,93,.29);
+$shadow-1--inset: inset 0px 0px 4px 0px rgba(0,0,0,0.37);
